requirements.txt
# Python 3.8 이상 필요 (권장: 3.9 이상)
# 자세한 버전 요구사항은 INSTALLATION.md 참고

# 데이터베이스 연결
pymongo>=4.6.0,<5.0.0

# 환경 변수 관리
python-dotenv>=1.0.0,<2.0.0

# 수치 연산 및 데이터 처리
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0

# 이미지 처리
Pillow>=10.0.0,<11.0.0
opencv-python>=4.8.0,<5.0.0

# 데이터 시각화
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0

# 머신러닝 유틸리티
scikit-learn>=1.3.0,<2.0.0
scikit-image>=0.21.0,<1.0.0

# 진행 상황 표시
tqdm>=4.65.0,<5.0.0

# 딥러닝 프레임워크
torch>=2.0.0
torchvision>=0.15.0

# 네트워크 통신 (연합학습용)
flask>=2.3.0
requests>=2.31.0

# Jupyter notebook
jupyter>=1.0.0
ipykernel>=6.25.0
